Move orchestration from S3 -> AWSSigV4 module
This seems like a reasonable organisation of the functionality needed to
pull together data for the S3 POST. Rather than having the S3 module
gathering data from various places, and storing a "simple" policy, the
bulk of this is now in the context's main module, `AWSSigV4`.

This module should act as the public interface, so this is good. The
Credential and S3 modules are both now very simple.