Add rough textarea functionality implementation
As of this, I have a textarea rendered, who's state is managed by Elm
(sorta. The browser also has a representation of this, no question...
It's a hard problem).

When a file is uploaded to S3, we parse the response, and append the
returned URL to the textarea's value.

This works!

Planned next steps:

* Render as markdown image, not just URL
* Tidy this. This Elm looks ugly to me.