Update markdown insertion to occur at cursor position
Now, rather than simply appending an image link to the textarea's
contents, we insert it at the current position of the cursor. This is
perhaps a little involved, due to the need to interact with a plain
javascript function outside of the Elm architecture, using Ports.

1. When a user selects a file with the rendered file input, we update
`model.fileToUpload` as before, but we also add in a command here:
`Ports.askPositionForLink`.
2. in js-land, we have a subscription set up for this port (in
app.js). When triggered, we use `selectionStart` to get the current
cursor position (this is not currently implemented in Elm).
3. Back in Elm-land, we also have a subscription set up, listening to
`Ports.receivePositionForLink`. This is triggered from js-land, with the
current cursor position.
4. Once this is called, we use UploadAsset as previously. When we get
down to the point of inserting the link, we use the new
`model.lastCursorPosition` to figure out where to put it.

This seems to work ok in some quick testing. One gotcha is that this
uses the actor model, the elm->js and js-elm calls are async, in theory
model state could change before we get a response back, and we'd end up
putting the wrong image in the wrong spot.

For now, I'm OK with this--I think the chances of it happening are slim.
If it seems to become an issue, I can revisit.