Use flexbox for layout
Flexbox is pretty well supported these days, so I think this is a good
choice.

This implementation gives me a full-height textarea, minus the button
(of unknown height!). The button sticks the the base of the window,
while the textarea's content can scroll. Very happy about this--try
doing that without flexbox!