Add *rough* end-to-end serviceworker implementation
This now sets up for sending messages between Elm, app.js, and otu
serviceworker, hopefully setting things up nicely for actually getting a
working offline setup.

As things stand, they are ugly, but work. When using the defer upload
button, a temporary markdown string should be set in the text area.
Later, once the file is uploaded, we should see this string be updated
to hold the real url.

This is in need of a refactor now that I have something workable, and
before I move onto the actual logic for getting it working offline.