Hook elm and serviceWorker retries together properly
With this commit the offline handling and retries on failure should be
working correctly.

* Rather than sending TESTHSH between elm and js, we
now send the idb id of the cached file, allowing identification later
on.
* The retries work by looking for a "status" field, which is also
persisted. Initially this value is not set, so we send a message from
the serviceWorker to its clients, which then talk to Elm through ports,
which attempts an upload.
  * There is no way for us to get a result back from this Elm work
  directly, so we always reject this first Promise, and schedule a
  retry.
  * Once we get into this retry we look up the row by id and check the
  status again. If succeeded, then we resolve the promise. Otherwise, we
  try the above again.
  * Over on the elm side, once an upload is successful we again send a
  message over ports to js, telling it so. Here we then update the row
  in idb, to show the correct status.

This is rather complicated, but I think it makes sense as perhaps the
best way to handle retries with synchronous work.